+++ public/locales/en/bundles.json
@@ -16,6 +16,7 @@
     },  
     "action": {
         "create_bundle": "Create new risk bundle",
-        "all_mine_bundles": "Show all risk bundles"
+        "all_mine_bundles": "Show all risk bundles",
+        "details": "Show bundle"
     }
 }
