+++ src/components/bundles/show_bundle.tsx
@@ -1,0 +1,9 @@
+import { BackendApi } from "../../backend/backend_api";
+
+interface ShowBundleProps {
+    backend: BackendApi;
+}
+
+export default function ShowBundle(props: ShowBundleProps) {
+    return (<></>);
+}
